requirements.txt
Flask == 2.0.1
Flask-Cors == 3.0.10
pymongo == 4.2.0
python-dotenv == 0.21.0
PyJWT == 2.5.0 
flask_jwt_extended == 4.4.4
dnspython == 2.2.1
pydantic == 1.10.2 
typing-extensions == 4.3.0
flask_pydantic == 0.11.0
motor == 3.0.0
orjson == 3.8.0